chore: refactor backchannel
