+++ src/modules/mod.rs
@@ -1,3 +1,4 @@
+pub mod backchannel;
 /// Utility functions for manipulating global values that can affect performance.
 pub mod globals;
 pub mod inner;
